refactoring for client-side
