+++ .vscode/settings.json
@@ -1,3 +1,26 @@
 {
-    "git.ignoreLimitWarning": true
+    "workbench.colorTheme": "Panda Syntax",
+    "workbench.iconTheme": "material-icon-theme",
+    "workbench.statusBar.visible": true,
+    "workbench.activityBar.visible": true,
+    "workbench.colorCustomizations": {
+        "statusBar.background": "#333333",
+        "statusBar.noFolderBackground": "#333333",
+        "statusBar.debuggingBackground": "#263238"
+    },
+    "editor.snippetSuggestions": "top",
+    "editor.fontSize": 13,
+    "editor.detectIndentation": false,
+    "editor.tabSize": 4,
+    "eslint.autoFixOnSave": true,
+    "eslint.enable": true,
+    "eslint.workingDirectories": [
+        "./client",
+        "./server"
+    ]
 }
+// "editorWarning.foreground": "#333333",
+// "editorError.foreground": "#333333",
+// "editorWarning.border": "#333333",
+// "editorError.border": "#333333",
+// "editor.minimap.enabled": false
