+++ client-side/.prettierrc.js
@@ -1,3 +1,3 @@
 module.exports = {
     "singleQuote": true
-};
+}
