+++ client-side/package.json
@@ -22,12 +22,6 @@
         "redux-promise": "^0.6.0",
         "redux-thunk": "^2.3.0"
     },
-    "scripts": {
-        "start": "react-scripts start",
-        "build": "react-scripts build",
-        "test": "react-scripts test --env=jsdom",
-        "eject": "react-scripts eject"
-    },
     "devDependencies": {
         "eslint": "^5.6.0",
         "eslint-config-airbnb": "^17.1.0",
@@ -37,5 +31,17 @@
         "eslint-plugin-prettier": "^2.6.2",
         "eslint-plugin-react": "^7.11.1",
         "prettier": "^1.14.3"
+    },
+    "scripts": {
+        "start": "react-scripts start",
+        "build": "react-scripts build",
+        "test": "react-scripts test --env=jsdom",
+        "eject": "react-scripts eject",
+        "lint": "eslint src/**/*.js --fix"
+    },
+    "proxy": {
+        "api": {
+            "target": "http://localhost:1111"
+        }
     }
 }
