+++ client-side/public/manifest.json
