+++ client-side/src/App.js
@@ -1,0 +1,19 @@
+import React, { Component } from 'react'
+import axios from 'axios'
+
+class App extends Component {
+
+    componentDidMount() {
+        axios.get('/api/product/brands').then(response => {
+            console.log(response)
+        })
+    }
+
+    render() {
+        return (
+            <div className='App'></div>
+        )
+    }
+}
+
+export default App
