--- client/src/App.js
@@ -1,14 +1,0 @@
-import React, {
-	Component
-} from 'react'
-import './App.css'
-
-class App extends Component {
-	render() {
-		return ( <
-			div className = 'App' > MY APP < /div>
-		)
-	}
-}
-
-export default App
