+++ server-side/.eslintrc.js
@@ -1,7 +1,10 @@
 module.exports = {
     "env": {
         "browser": true,
-        "commonjs": true
+        "node": true
+    },
+    "parserOptions": {
+        "ecmaVersion": 6
     },
     "extends": "eslint:recommended",
     "rules": {
