+++ server-side/server/middleware/admin.js
@@ -1,0 +1,8 @@
+let admin = (req,res,next) =>{
+	if(req.user.role === 0 ){
+		return res.send('you are not allowed, get out now.')
+	}
+	next()
+}
+
+module.exports = { admin }
