+++ server-side/server/models/brand.js
@@ -1,0 +1,14 @@
+const mongoose = require('mongoose')
+
+const brandSchema = mongoose.Schema({
+	name:{
+		required: true,
+		type: String,
+		unique: 1,
+		maxlength:100
+	}
+})
+
+const Brand = mongoose.model('Brand',brandSchema)
+
+module.exports = { Brand }
