+++ server-side/server/models/user.js
@@ -84,10 +84,7 @@ userSchema.statics.findByToken = function (token, cb) {
 	var user = this
 
 	jwt.verify(token, process.env.SECRET, function (err, decode) {
-		user.findOne({
-			'_id': decode,
-			'token': token
-		}, function (err, user) {
+		user.findOne({ '_id': decode, 'token': token }, function (err, user) {
 			if (err) return cb(err)
 			cb(null, user)
 		})
@@ -98,6 +95,4 @@ userSchema.statics.findByToken = function (token, cb) {
 
 const User = mongoose.model('User', userSchema)
 
-module.exports = {
-	User
-}
+module.exports = { User }
