--- server/middleware/admin.js
@@ -1,10 +1,0 @@
-let admin = (req, res, next) => {
-	if (req.user.role === 0) {
-		return res.send('you are not authorized')
-	}
-	next()
-}
-
-module.exports = {
-	admin
-}
