--- server/models/wood.js
@@ -1,16 +1,0 @@
-const mongoose = require('mongoose')
-
-const woodSchema = mongoose.Schema({
-	name: {
-		required: true,
-		type: String,
-		unique: 1,
-		maxlength: 100
-	}
-})
-
-const Wood = mongoose.model('Wood', woodSchema)
-
-module.exports = {
-	Wood
-}
